refactor: rename post-page folder into entry-page
