+++ ui/apps/akasha/src/components/entry-page/common/base-entry-page.tsx
