+++ ui/apps/akasha/src/components/entry-page/common/pending-reply.tsx
