+++ ui/apps/akasha/src/components/entry-page/post-page.tsx
